- fixed SCF_DEBUG to work correctly with embeded interfaces         - fixed some compiler warnings         - fixed definition of int32 and uint32 should be long not int         - removed the type UByte,SByte, UShort,SShort, ULong,SLong, UInt,           SInt, uchar, ushort and ulong and substituted code with the           corresponding intxx types.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@9122 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csengine/engine.h
@@ -759,7 +759,7 @@ public:
    * array.
    */
   virtual int GetNearbyLights (iSector* sector, const csVector3& pos,
-  	ULong flags, iLight** lights, int max_num_lights);
+  	uint32 flags, iLight** lights, int max_num_lights);
 
   /**
    * This routine returns an iterator to iterate over
