+++ plugins/mesh/impexp/3ds/ie3ds.cpp
@@ -244,7 +244,7 @@ static void AssignDefinedTexels (iModelDataObject * pDataObject,
   }
 }
 
-iModelData *csModelConverter3ds::Load( UByte* buffer, ULong size )
+iModelData *csModelConverter3ds::Load( uint8* buffer, uint32 size )
 {
   Lib3dsFile *p3dsFile;
   csModelData *pModelData;
@@ -471,7 +471,7 @@ bool csModelConverter3ds::LoadMeshObjectData( iModelDataObject *pDataObject,
   return true;
 }
 
-Lib3dsFile *csModelConverter3ds::LoadFileData( UByte* pBuffer, ULong size )
+Lib3dsFile *csModelConverter3ds::LoadFileData( uint8* pBuffer, uint32 size )
 {
   // This code is pulled from lib3ds
   Lib3dsFile *pFile;
